refactor(cli): give the async caching actors a single home
`AsyncThrowableCaching` was private to SwiftVersionProvider.swift, so the
SDK deployment target provider grew a near-identical non-throwing copy.
Both now live in TuistSupport/Utils/AsyncCaching.swift and both providers
use them from there.

They stay two actors rather than one generic over the failure type
because `Task` is only constructible with `Never` or `any Error`; typed
throws never generalized it. The file says so, so the next reader doesn't
have to rediscover it.
